app_requirements.txt
streamlit==1.30.0
faiss-cpu==1.7.4
sentence-transformers==2.3.1
pandas==2.1.4
numpy==1.26.3
pillow==10.2.0
requests==2.31.0 